+++ app/Http/Controllers/Admin/GalleryController.php
@@ -4,11 +4,12 @@ namespace App\Http\Controllers\Admin;
 
 use App\Models\Gallery;
 use App\Models\Property;
-use Illuminate\Http\Request;
 use Illuminate\Contracts\View\View;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\RedirectResponse;
+use Symfony\Component\HttpFoundation\Response;
 use App\Http\Requests\Admin\ValidateGalleryRequest;
 
 class GalleryController extends Controller
@@ -28,6 +29,11 @@ class GalleryController extends Controller
 
     public function edit(Property $property,Gallery $gallery): View
     {
+        abort_if(Gate::denies('property_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        if($property->agent->name != auth()->user()->name && auth()->user()->roles()->where('title', 'agent')->count() > 0){
+            abort(403);
+         }
+         
         return view('admin.galleries.edit', compact('gallery', 'property'));
     }
 
@@ -47,6 +53,11 @@ class GalleryController extends Controller
 
     public function destroy(Property $property,Gallery $gallery): RedirectResponse
     {
+        abort_if(Gate::denies('property_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        if($property->agent->name != auth()->user()->name && auth()->user()->roles()->where('title', 'agent')->count() > 0){
+            abort(403);
+         }
+
         if($gallery->path){
             File::delete('storage/' . $gallery->path);
         }
