+++ app/Http/Controllers/Admin/UserController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers\Admin;
 use App\Models\Role;
 use App\Models\User;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Request;
 use App\Http\Requests\Admin\StoreUserRequest;
 use App\Http\Requests\Admin\UpdateUserRequest;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
@@ -18,6 +20,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('user_management_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $users = User::all();
 
         return view('admin.users.index', compact('users'));
@@ -30,6 +34,7 @@ class UserController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('user_management_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $roles = Role::pluck('title', 'id');
 
         return view('admin.users.create', compact('roles'));
@@ -43,6 +48,7 @@ class UserController extends Controller
      */
     public function store(StoreUserRequest $request)
     {
+        abort_if(Gate::denies('user_management_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $user = User::create($request->validated() + ['password' => bcrypt($request->password)]);
         $user->roles()->sync($request->input('roles'));
 
@@ -60,6 +66,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        abort_if(Gate::denies('user_management_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $roles = Role::pluck('title', 'id');
 
         return view('admin.users.edit', compact('user','roles'));
@@ -74,6 +81,7 @@ class UserController extends Controller
      */
     public function update(UpdateUserRequest $request,User $user)
     {
+        abort_if(Gate::denies('user_management_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $user->update($request->validated() + ['password' => bcrypt($request->password)]);
         $user->roles()->sync($request->input('roles'));
 
@@ -91,6 +99,7 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        abort_if(Gate::denies('user_management_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         $user->delete();
 
         return redirect()->route('admin.users.index')->with([
